Reworked code to use TCP
This is part of a larger refactor to use an updated protocol (which also
switched to TCP).
ListenerServer and BroadcastServer are now merged into one as this server
will now use Twisted for underlying network comms instead of SocketServer.
This new Class is called CommunicationServer, which will expose functions
for sending and recieving strings.